feat: move legacy Popover to the new entrypoint
