--- src/components/Popover/Popover.classname.ts
@@ -1,3 +1,0 @@
-import {block} from '../utils/cn';
-
-export const cnPopover = block('popover');
