--- src/components/Popover/__stories__/PopoverDemo.classname.ts
@@ -1,3 +1,0 @@
-import {block} from '../../utils/cn';
-
-export const cnPopoverDemo = block('popover-demo');
