+++ src/components/legacy/Popover/Popover.classname.ts
@@ -1,0 +1,3 @@
+import {block} from '../../utils/cn';
+
+export const cnPopover = block('popover-legacy');
