+++ src/components/legacy/Popover/__stories__/PopoverDemo.classname.ts
@@ -1,0 +1,3 @@
+import {block} from '../../../utils/cn';
+
+export const cnPopoverDemo = block('popover-legacy-demo');
