+++ src/components/legacy/Popover/components/Links/Links.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import {Link} from '../../../Link';
+import {Link} from '../../../../Link';
 import {cnPopover} from '../../Popover.classname';
 
 export type LinksProps = {
