+++ src/legacy.ts
@@ -1,0 +1,1 @@
+export * from './components/legacy';
